Large update

+++ Loja.Importer/Importador.cs
@@ -4,11 +4,16 @@ using System.Linq;
 using System.Text;
 using System.IO;
 using System.Threading.Tasks;
-using Loja.Modelos;
 using Loja.Data;
+using Loja.Models;
 
 namespace Loja.Importer
 {
+    /// <summary>
+    /// 
+    /// Sumário serve para expor informação adicional aos comentários ou para informar todos os bugs na aplicação em causa.
+    /// 
+    /// </summary>
     public class Importador
     {
         public void Importer()
@@ -33,7 +38,7 @@ namespace Loja.Importer
                         var substrings = product.Split(';');
                         produto.Id = substrings[0];
                         produto.Nome = substrings[1];
-                        produto.Preco = Convert.ToDecimal(substrings[2]);
+                        produto.Preco = Convert.ToDouble(substrings[2]);
                         produto.Tipo = substrings[3];
                         produto.DataDeAquisicao = DateTime.Parse(substrings[4]);
                         produto.DataDeVenda = DateTime.Parse(substrings[5]);
